Fix the instruction size to 16 bit, and modify the gpio test case to use tick()

+++ src/main.cpp
@@ -2,7 +2,7 @@
 #include <cstdint>
 #include "PioStateMachine.h"
 
-using u32 = uint32_t;
+using u16 = uint32_t;
 int main()
 {
     PioStateMachine pio0;
